cmd/keyfile: improve ergonomics of keyfile creation
Don't attempt to create the file till after the passphrase and keyring
generation has succeeded, so that we don't leave around an empty file on error.
Add an early check for file existence so we can exit early. We'll still check
(again) at file creation for non-existence, but this makes common mistakes less
likely to be frictional.